requirements.txt
beautifulsoup4
frontend
google-cloud-texttospeech
llm
llm-cmd
nltk
openai
sentence-transformers
thefuzz
transformers
wget
PyPDF2
tools
spacy
tiktoken
pyperclip 
PyDrive2
PyMuPDF
ffprobe-python
ffmpeg-python
fastapi
